- added SkipWhileInclusive - added branching from sequences - updated methods in StringX

+++ core/src/test/java/hzt/collections/MapXTest.java
@@ -47,7 +47,7 @@ class MapXTest {
                 .mapToInt(LocalDate::getDayOfMonth)
                 .sum();
 
-        final int actual = MapX.of(museumMap)
+        final long actual = MapX.of(museumMap)
                 .valuesToIterX(Museum::getDateOfOpening)
                 .sumOfInts(LocalDate::getDayOfMonth);
